Split tests into separate unit and integration test classes
