--- src/Tests/Eshopworld.DevOps.Tests/TestKey2
@@ -1,1 +1,0 @@
-﻿testVal2
